File renaming for test components
